+++ src/index.ts
@@ -1,4 +1,4 @@
-import { MarkdownView, Plugin } from "obsidian";
+import { Plugin } from "obsidian";
 import { ObsidianOutlinerPluginSettingTab, Settings } from "./settings";
 import { IFeature } from "./feature";
 import { ObsidianUtils } from "./obsidian_utils";
@@ -13,10 +13,7 @@ import { MoveCursorToPreviousUnfoldedLineFeature } from "./features/MoveCursorTo
 import { EnsureCursorInListContentFeature } from "./features/EnsureCursorInListContentFeature";
 import { DeleteShouldIgnoreBulletsFeature } from "./features/DeleteShouldIgnoreBulletsFeature";
 import { SelectionShouldIgnoreBulletsFeature } from "./features/SelectionShouldIgnoreBulletsFeature";
-
-class ZoomState {
-  constructor(public line: CodeMirror.LineHandle, public header: HTMLElement) {}
-}
+import { ZoomFeature } from "./features/ZoomFeature";
 
 export default class ObsidianOutlinerPlugin extends Plugin {
   private features: IFeature[];
@@ -25,7 +22,6 @@ export default class ObsidianOutlinerPlugin extends Plugin {
   private obsidianUtils: ObsidianUtils;
   private editorUtils: EditorUtils;
   private listsUtils: ListUtils;
-  private zoomStates: WeakMap<CodeMirror.Editor, ZoomState> = new WeakMap();
 
   execute(editor: CodeMirror.Editor, cb: (root: Root) => boolean): boolean {
     const root = this.listsUtils.parseList(editor, editor.getCursor());
@@ -77,116 +73,6 @@ export default class ObsidianOutlinerPlugin extends Plugin {
     return this.setFold(editor, "unfold");
   }
 
-  zoomOut(editor: CodeMirror.Editor) {
-    const zoomState = this.zoomStates.get(editor);
-
-    if (!zoomState) {
-      return false;
-    }
-
-    for (let i = editor.firstLine(), l = editor.lastLine(); i <= l; i++) {
-      editor.removeLineClass(i, "wrap", "outliner-plugin-hidden-row");
-    }
-
-    zoomState.header.parentElement.removeChild(zoomState.header);
-
-    this.zoomStates.delete(editor);
-
-    return true;
-  }
-
-  zoomIn(
-    editor: CodeMirror.Editor,
-    cursor: CodeMirror.Position = editor.getCursor()
-  ) {
-    const lineNo = cursor.line;
-    const root = this.listsUtils.parseList(editor, cursor);
-
-    if (!root) {
-      return false;
-    }
-
-    this.zoomOut(editor);
-
-    const { indentLevel } = this.listsUtils.getListLineInfo(
-      editor.getLine(lineNo),
-      root.getIndentSign()
-    );
-
-    let after = false;
-    for (let i = editor.firstLine(), l = editor.lastLine(); i <= l; i++) {
-      if (i < lineNo) {
-        editor.addLineClass(i, "wrap", "outliner-plugin-hidden-row");
-      } else if (i > lineNo && !after) {
-        const afterLineInfo = this.listsUtils.getListLineInfo(
-          editor.getLine(i),
-          root.getIndentSign()
-        );
-        after = !afterLineInfo || afterLineInfo.indentLevel <= indentLevel;
-      }
-
-      if (after) {
-        editor.addLineClass(i, "wrap", "outliner-plugin-hidden-row");
-      }
-    }
-
-    const createSeparator = () => {
-      const span = document.createElement("span");
-      span.textContent = " > ";
-      return span;
-    };
-
-    const createTitle = (content: string, cb: () => void) => {
-      const a = document.createElement("a");
-      a.className = "outliner-plugin-zoom-title";
-      if (content) {
-        a.textContent = content;
-      } else {
-        a.innerHTML = "&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;";
-      }
-      a.onclick = (e) => {
-        e.preventDefault();
-        cb();
-      };
-      return a;
-    };
-
-    const createHeader = () => {
-      const div = document.createElement("div");
-      div.className = "outliner-plugin-zoom-header";
-
-      let list = root.getListUnderLine(lineNo).getParent();
-      while (list && list.getParent()) {
-        const lineNo = root.getLineNumberOf(list);
-        div.prepend(
-          createTitle(list.getContent(), () =>
-            this.zoomIn(editor, { line: lineNo, ch: 0 })
-          )
-        );
-        div.prepend(createSeparator());
-        list = list.getParent();
-      }
-
-      div.prepend(
-        createTitle(this.app.workspace.activeLeaf.getDisplayText(), () =>
-          this.zoomOut(editor)
-        )
-      );
-
-      return div;
-    };
-
-    const zoomHeader = createHeader();
-    editor.getWrapperElement().prepend(zoomHeader);
-
-    this.zoomStates.set(
-      editor,
-      new ZoomState(editor.getLineHandle(lineNo), zoomHeader)
-    );
-
-    return true;
-  }
-
   selectAll(editor: CodeMirror.Editor) {
     const selections = editor.listSelections();
 
@@ -231,24 +117,6 @@ export default class ObsidianOutlinerPlugin extends Plugin {
     }
 
     return true;
-  }
-
-  createCommandCallback(cb: (editor: CodeMirror.Editor) => boolean) {
-    return () => {
-      const view = this.app.workspace.getActiveViewOfType(MarkdownView);
-
-      if (!view) {
-        return;
-      }
-
-      const editor = view.sourceMode.cmEditor;
-
-      const worked = cb(editor);
-
-      if (!worked && window.event && window.event.type === "keydown") {
-        (editor as any).triggerOnKeyDown(window.event);
-      }
-    };
   }
 
   async onload() {
@@ -302,6 +170,7 @@ export default class ObsidianOutlinerPlugin extends Plugin {
         this.settings,
         this.listsUtils
       ),
+      new ZoomFeature(this, this.obsidianUtils, this.listsUtils),
     ];
 
     for (const feature of this.features) {
@@ -311,7 +180,9 @@ export default class ObsidianOutlinerPlugin extends Plugin {
     this.addCommand({
       id: "move-list-item-up",
       name: "Move list item up",
-      callback: this.createCommandCallback(this.moveListElementUp.bind(this)),
+      callback: this.obsidianUtils.createCommandCallback(
+        this.moveListElementUp.bind(this)
+      ),
       hotkeys: [
         {
           modifiers: ["Mod", "Shift"],
@@ -323,7 +194,9 @@ export default class ObsidianOutlinerPlugin extends Plugin {
     this.addCommand({
       id: "move-list-item-down",
       name: "Move list item down",
-      callback: this.createCommandCallback(this.moveListElementDown.bind(this)),
+      callback: this.obsidianUtils.createCommandCallback(
+        this.moveListElementDown.bind(this)
+      ),
       hotkeys: [
         {
           modifiers: ["Mod", "Shift"],
@@ -335,7 +208,7 @@ export default class ObsidianOutlinerPlugin extends Plugin {
     this.addCommand({
       id: "indent-list",
       name: "Indent list",
-      callback: this.createCommandCallback(
+      callback: this.obsidianUtils.createCommandCallback(
         this.moveListElementRight.bind(this)
       ),
       hotkeys: [
@@ -349,7 +222,9 @@ export default class ObsidianOutlinerPlugin extends Plugin {
     this.addCommand({
       id: "select-all",
       name: "Select list item or whole list",
-      callback: this.createCommandCallback(this.selectAll.bind(this)),
+      callback: this.obsidianUtils.createCommandCallback(
+        this.selectAll.bind(this)
+      ),
       hotkeys: [
         {
           modifiers: ["Mod"],
@@ -361,7 +236,7 @@ export default class ObsidianOutlinerPlugin extends Plugin {
     this.addCommand({
       id: "fold",
       name: "Fold list",
-      callback: this.createCommandCallback(this.fold.bind(this)),
+      callback: this.obsidianUtils.createCommandCallback(this.fold.bind(this)),
       hotkeys: [
         {
           modifiers: ["Mod"],
@@ -373,7 +248,9 @@ export default class ObsidianOutlinerPlugin extends Plugin {
     this.addCommand({
       id: "unfold",
       name: "Unfold list",
-      callback: this.createCommandCallback(this.unfold.bind(this)),
+      callback: this.obsidianUtils.createCommandCallback(
+        this.unfold.bind(this)
+      ),
       hotkeys: [
         {
           modifiers: ["Mod"],
@@ -385,130 +262,15 @@ export default class ObsidianOutlinerPlugin extends Plugin {
     this.addCommand({
       id: "outdent-list",
       name: "Outdent list",
-      callback: this.createCommandCallback(this.moveListElementLeft.bind(this)),
+      callback: this.obsidianUtils.createCommandCallback(
+        this.moveListElementLeft.bind(this)
+      ),
       hotkeys: [
         {
           modifiers: ["Shift"],
           key: "Tab",
         },
       ],
-    });
-
-    this.addCommand({
-      id: "zoom-in",
-      name: "Zoom In",
-      callback: this.createCommandCallback(this.zoomIn.bind(this)),
-      hotkeys: [
-        {
-          modifiers: ["Mod"],
-          key: ".",
-        },
-      ],
-    });
-
-    this.addCommand({
-      id: "zoom-out",
-      name: "Zoom Out",
-      callback: this.createCommandCallback(this.zoomOut.bind(this)),
-      hotkeys: [
-        {
-          modifiers: ["Mod", "Shift"],
-          key: ".",
-        },
-      ],
-    });
-
-    this.registerCodeMirror((cm) => {
-      this.attachZoomModeHandlers(cm);
-    });
-  }
-
-  attachZoomModeHandlers(cm: CodeMirror.Editor) {
-    cm.on("beforeChange", (cm, changeObj) => {
-      const zoomState = this.zoomStates.get(cm);
-
-      if (
-        !zoomState ||
-        changeObj.origin !== "setValue" ||
-        changeObj.from.line !== 0 ||
-        changeObj.from.ch !== 0
-      ) {
-        return;
-      }
-
-      const tillLine = cm.lastLine();
-      const tillCh = cm.getLine(tillLine).length;
-
-      if (changeObj.to.line !== tillLine || changeObj.to.ch !== tillCh) {
-        return;
-      }
-
-      this.zoomOut(cm);
-    });
-
-    cm.on("change", (cm, changeObj) => {
-      const zoomState = this.zoomStates.get(cm);
-
-      if (!zoomState || changeObj.origin !== "setValue") {
-        return;
-      }
-
-      this.zoomIn(cm, {
-        line: cm.getLineNumber(zoomState.line),
-        ch: 0,
-      });
-    });
-
-    cm.on("beforeSelectionChange", (cm, changeObj) => {
-      if (!this.zoomStates.has(cm)) {
-        return;
-      }
-
-      let visibleFrom: CodeMirror.Position | null = null;
-      let visibleTill: CodeMirror.Position | null = null;
-
-      for (let i = cm.firstLine(); i <= cm.lastLine(); i++) {
-        const wrapClass = cm.lineInfo(i).wrapClass || "";
-        const isHidden = wrapClass.includes("outliner-plugin-hidden-row");
-        if (visibleFrom === null && !isHidden) {
-          visibleFrom = { line: i, ch: 0 };
-        }
-        if (visibleFrom !== null && visibleTill !== null && isHidden) {
-          break;
-        }
-        if (visibleFrom !== null) {
-          visibleTill = { line: i, ch: cm.getLine(i).length };
-        }
-      }
-
-      let changed = false;
-
-      for (const range of changeObj.ranges) {
-        if (range.anchor.line < visibleFrom.line) {
-          changed = true;
-          range.anchor.line = visibleFrom.line;
-          range.anchor.ch = visibleFrom.ch;
-        }
-        if (range.anchor.line > visibleTill.line) {
-          changed = true;
-          range.anchor.line = visibleTill.line;
-          range.anchor.ch = visibleTill.ch;
-        }
-        if (range.head.line < visibleFrom.line) {
-          changed = true;
-          range.head.line = visibleFrom.line;
-          range.head.ch = visibleFrom.ch;
-        }
-        if (range.head.line > visibleTill.line) {
-          changed = true;
-          range.head.line = visibleTill.line;
-          range.head.ch = visibleTill.ch;
-        }
-      }
-
-      if (changed) {
-        changeObj.update(changeObj.ranges);
-      }
     });
   }
