+++ src/obsidian_utils.ts
@@ -1,4 +1,4 @@
-import { App } from "obsidian";
+import { App, MarkdownView } from "obsidian";
 
 export interface IObsidianTabsSettigns {
   useTab: boolean;
@@ -13,6 +13,28 @@ export class ObsidianUtils {
       useTab: true,
       tabSize: 4,
       ...(this.app.vault as any).config,
+    };
+  }
+
+  getActiveLeafDisplayText() {
+    return this.app.workspace.activeLeaf.getDisplayText();
+  }
+
+  createCommandCallback(cb: (editor: CodeMirror.Editor) => boolean) {
+    return () => {
+      const view = this.app.workspace.getActiveViewOfType(MarkdownView);
+
+      if (!view) {
+        return;
+      }
+
+      const editor = view.sourceMode.cmEditor;
+
+      const worked = cb(editor);
+
+      if (!worked && window.event && window.event.type === "keydown") {
+        (editor as any).triggerOnKeyDown(window.event);
+      }
     };
   }
 }
